requirements.txt
asgiref==3.9.2
Django==5.2.7
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
